fix double link ~/.stumpwmrc.d

+++ scripts/link-inits.sh
@@ -26,6 +26,7 @@ ln -sf "${INITS_TOP}/keynavs/.keynavrc.mine" "${HOME}/.keynavrc"
 ln -sf "${STUMPWM_TOP}/lisp/.stumpwmrc" "${HOME}/"
 
 #link the top itself
-ln -sf "${STUMPWM_TOP}" "${HOME}/.stumpwmrc.d"
+#if existss and is a directory, link will be put inside the directory symlink
+test ! -e "${HOME}/.stumpwmrc.d" && ln -sf "${STUMPWM_TOP}" "${HOME}/.stumpwmrc.d"
 ln -sf "${HOME}/repos/dotemacs" "${HOME}/"
